+++ Grbl_Esp32/src/Planner.cpp
@@ -252,10 +252,10 @@ uint8_t plan_check_full_buffer() {
 float plan_compute_profile_nominal_speed(plan_block_t* block) {
     float nominal_speed = block->programmed_rate;
     if (block->motion.rapidMotion) {
-        nominal_speed *= (0.01 * sys.r_override);
+        nominal_speed *= (0.01f * sys.r_override);
     } else {
         if (!(block->motion.noFeedOverride)) {
-            nominal_speed *= (0.01 * sys.f_override);
+            nominal_speed *= (0.01f * sys.f_override);
         }
         if (nominal_speed > block->rapid_rate) {
             nominal_speed = block->rapid_rate;
@@ -400,10 +400,10 @@ bool plan_buffer_line(float* target, plan_line_data_t* pl_data) {
             } else {
                 convert_delta_vector_to_unit_vector(junction_unit_vec);
                 float junction_acceleration = limit_acceleration_by_axis_maximum(junction_unit_vec);
-                float sin_theta_d2          = sqrt(0.5 * (1.0 - junction_cos_theta));  // Trig half angle identity. Always positive.
+                float sin_theta_d2          = float(sqrt(0.5f * (1.0f - junction_cos_theta)));  // Trig half angle identity. Always positive.
                 block->max_junction_speed_sqr =
                     MAX(MINIMUM_JUNCTION_SPEED * MINIMUM_JUNCTION_SPEED,
-                        (junction_acceleration * config->_junctionDeviation * sin_theta_d2) / (1.0 - sin_theta_d2));
+                        (junction_acceleration * config->_junctionDeviation * sin_theta_d2) / (1.0f - sin_theta_d2));
             }
         }
     }
